finshed comment API

+++ src/main/java/com/zpc/community/controller/PublishController.java
@@ -27,7 +27,7 @@ public class PublishController {
     @Autowired
     private QuestionService questionService;
     @GetMapping("/publish/{id}")
-    public String edit(@PathVariable(name = "id") Integer id,
+    public String edit(@PathVariable(name = "id") Long id,
                        Model model){
         QuestionDTO quesstion = questionService.getById(id);
         model.addAttribute("title", quesstion.getTitle());
@@ -47,7 +47,7 @@ public class PublishController {
             @RequestParam(value = "title",required = false) String title,
             @RequestParam(value ="description",required = false) String description,
             @RequestParam(value ="tag",required = false) String tag,
-            @RequestParam(value ="id") Integer id,
+            @RequestParam(value ="id") Long id,
             HttpServletRequest request,
             Model model
     ){
