+++ src/main/java/com/zpc/community/controller/QuestionController.java
@@ -24,7 +24,7 @@ public class QuestionController {
 
     @GetMapping("/question/{id}")
     public String question(
-            @PathVariable(name = "id") Integer id,
+            @PathVariable(name = "id") Long id,
             Model model
     ){
         QuestionDTO questionDTO = questionService.getById(id);
